+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import baseball.domain.Controller.Controller;
+import baseball.Controller.Controller;
 
 public class Application {
     public static void main(String[] args) {
