+++ src/main/java/baseball/domain/GameResult.java
@@ -1,6 +1,6 @@
 package baseball.domain;
 
-import baseball.domain.dto.GameResultResponseDto;
+import baseball.dto.GameResultResponseDto;
 
 import java.util.Objects;
