+++ src/test/java/baseball/domain/GameResultTest.java
@@ -1,6 +1,6 @@
 package baseball.domain;
 
-import baseball.domain.dto.GameResultResponseDto;
+import baseball.dto.GameResultResponseDto;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
